create test_base.py for common base tests
